First pass at fulfill mutation
