+++ src/data/exchange.graphql
@@ -36,22 +36,63 @@ type CreateOrderWithArtworkPayload {
   order: Order
 }
 
-# A valid URL, transported as a string
+# Date in YYYY-MM-DD format
+scalar Date
+
+# DateTime
 scalar DateTime
 
-# Autogenerated input type of FinalizeOrder
-input FinalizeOrderInput {
+# Autogenerated input type of FulfillWithOneFulfillment
+input FulfillWithOneFulfillmentInput {
   # A unique identifier for the client performing the mutation.
   clientMutationId: String
+  fulfillment: FulfillmentAttributes!
   id: ID!
 }
 
-# Autogenerated return type of FinalizeOrder
-type FinalizeOrderPayload {
+# Autogenerated return type of FulfillWithOneFulfillment
+type FulfillWithOneFulfillmentPayload {
   # A unique identifier for the client performing the mutation.
   clientMutationId: String
   errors: [String!]!
   order: Order
+}
+
+# A Fulfillment for an order
+type Fulfillment {
+  courier: String!
+  createdAt: DateTime!
+  estimatedDelivery: Date
+  id: ID!
+  notes: String
+  trackingId: String
+  updatedAt: DateTime!
+}
+
+# Attributes of a Fulfillment
+input FulfillmentAttributes {
+  courier: String!
+  estimatedDelivery: Date
+  notes: String
+  trackingId: String
+}
+
+# The connection type for Fulfillment.
+type FulfillmentConnection {
+  # A list of edges.
+  edges: [FulfillmentEdge]
+
+  # Information to aid in pagination.
+  pageInfo: PageInfo!
+}
+
+# An edge in a connection.
+type FulfillmentEdge {
+  # A cursor for use in pagination.
+  cursor: String!
+
+  # The item at the end of the edge.
+  node: Fulfillment
 }
 
 # A Line Item
@@ -59,6 +100,19 @@ type LineItem {
   artworkId: String!
   createdAt: DateTime!
   editionSetId: String
+  fulfillments(
+    # Returns the elements in the list that come after the specified cursor.
+    after: String
+
+    # Returns the elements in the list that come before the specified cursor.
+    before: String
+
+    # Returns the first _n_ elements from the list.
+    first: Int
+
+    # Returns the last _n_ elements from the list.
+    last: Int
+  ): FulfillmentConnection
   id: ID!
   priceCents: Int!
   quantity: Int!
@@ -86,7 +140,9 @@ type LineItemEdge {
 type Mutation {
   approveOrder(input: ApproveOrderInput!): ApproveOrderPayload
   createOrderWithArtwork(input: CreateOrderWithArtworkInput!): CreateOrderWithArtworkPayload
-  finalizeOrder(input: FinalizeOrderInput!): FinalizeOrderPayload
+
+  # Fulfill an order with one Fulfillment, it sets this fulfillment to each line item in order
+  fulfillWithOneFulfillment(input: FulfillWithOneFulfillmentInput!): FulfillWithOneFulfillmentPayload
   rejectOrder(input: RejectOrderInput!): RejectOrderPayload
   setPayment(input: SetPaymentInput!): SetPaymentPayload
   setShipping(input: SetShippingInput!): SetShippingPayload
@@ -99,6 +155,7 @@ type Order {
   code: String!
   commissionFeeCents: Int
   createdAt: DateTime!
+  creditCardId: String
   currencyCode: String!
   fulfillmentType: OrderFulfillmentTypeEnum
   id: ID!
@@ -173,8 +230,8 @@ enum OrderStateEnum {
   # order is approved by partner
   APPROVED
 
-  # order is finalized by partner
-  FINALIZED
+  # order is fulfilled by partner
+  FULFILLED
 
   # order is still pending submission by collector
   PENDING
