--- src/schema/ecommerce/finalize_order_mutation.js
@@ -1,82 +1,0 @@
-import { graphql } from "graphql"
-import { OrderReturnType } from "schema/ecommerce/types/order_return"
-import { OrderMutationInputType } from "schema/ecommerce/types/order_mutation_input"
-import { mutationWithClientMutationId } from "graphql-relay"
-
-export const FinalizeOrderMutation = mutationWithClientMutationId({
-  name: "FinalizeOrder",
-  description: "Finalizes an order",
-  inputFields: OrderMutationInputType.getFields(),
-  outputFields: {
-    result: {
-      type: OrderReturnType,
-      resolve: order => order,
-    },
-  },
-  mutateAndGetPayload: (
-    { orderId, creditCardId },
-    context,
-    { rootValue: { accessToken, exchangeSchema } }
-  ) => {
-    if (!accessToken) {
-      return new Error("You need to be signed in to perform this action")
-    }
-
-    const mutation = `
-      mutation finalizeOrder($orderId: ID!) {
-        ecommerce_finalizeOrder(input: {
-          id: $orderId,
-        }) {
-          order {
-           id
-            code
-            currencyCode
-            state
-            partnerId
-            userId
-            fulfillmentType
-            shippingAddressLine1
-            shippingAddressLine2
-            shippingCity
-            shippingCountry
-            shippingPostalCode
-            shippingRegion
-            itemsTotalCents
-            shippingTotalCents
-            taxTotalCents
-            commissionFeeCents
-            transactionFeeCents
-            buyerTotalCents
-            sellerTotalCents
-            updatedAt
-            createdAt
-            stateUpdatedAt
-            stateExpiresAt
-            lineItems{
-              edges{
-                node{
-                  id
-                  priceCents
-                  artworkId
-                  editionSetId
-                  quantity
-                }
-              }
-            }
-          }
-          errors
-        }
-      }
-    `
-    return graphql(exchangeSchema, mutation, null, context, {
-      orderId,
-      creditCardId,
-    }).then(result => {
-      const { order, errors } = result.data.ecommerce_finalizeOrder
-      return {
-        order,
-        errors,
-      }
-    })
-  },
-})
