+++ src/schema/ecommerce/fulfill_order_with_one_fulfillment_mutation.js
@@ -1,0 +1,125 @@
+import {
+  GraphQLInputObjectType,
+  GraphQLNonNull,
+  GraphQLString,
+  graphql,
+} from "graphql"
+import { OrderReturnType } from "schema/ecommerce/types/order_return"
+import { mutationWithClientMutationId } from "graphql-relay"
+
+const FulfillmentInputType = new GraphQLInputObjectType({
+  name: "FulfillmentInputType",
+  fields: {
+    courier: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: "Courier of the fulfiller",
+    },
+    trackingId: {
+      type: GraphQLString,
+      description: "Courier's Tracking ID of this fulfillment",
+    },
+    estimatedDelivery: {
+      type: GraphQLString,
+      description: "Estimated delivery in YY-MM-DD format",
+    },
+  },
+})
+
+const FulfillOrderWithOneFulfillmentInputType = new GraphQLInputObjectType({
+  name: "FulfillOrderWithOneFulfillmentInput",
+  fields: {
+    orderId: {
+      type: new GraphQLNonNull(GraphQLString),
+      description: "ID of the order",
+    },
+    fulfillment: {
+      type: GraphQLNonNull(FulfillmentInputType),
+      description: "Fulfillment information of this order",
+    },
+  },
+})
+
+export const FulfillOrderWithOneFulfillmentMutation = mutationWithClientMutationId(
+  {
+    name: "FulfillOrderWithOneFulfillment",
+    description:
+      "Fulfills an Order with one fulfillment by setting this fulfillment to all line items of this order",
+    inputFields: FulfillOrderWithOneFulfillmentInputType.getFields(),
+    outputFields: {
+      result: {
+        type: OrderReturnType,
+        resolve: order => order,
+      },
+    },
+    mutateAndGetPayload: (
+      { orderId, fulfillment },
+      context,
+      { rootValue: { accessToken, exchangeSchema } }
+    ) => {
+      if (!accessToken) {
+        return new Error("You need to be signed in to perform this action")
+      }
+
+      const mutation = `
+      mutation finalizeOrder($orderId: ID!) {
+        ecommerce_fulfillWithOneFulfillment(input: {
+          id: $orderId,
+          fulfillment: $fulfillment
+        }) {
+          order {
+           id
+            code
+            currencyCode
+            state
+            partnerId
+            userId
+            fulfillmentType
+            shippingAddressLine1
+            shippingAddressLine2
+            shippingCity
+            shippingCountry
+            shippingPostalCode
+            shippingRegion
+            itemsTotalCents
+            shippingTotalCents
+            taxTotalCents
+            commissionFeeCents
+            transactionFeeCents
+            buyerTotalCents
+            sellerTotalCents
+            updatedAt
+            createdAt
+            stateUpdatedAt
+            stateExpiresAt
+            lineItems{
+              edges{
+                node{
+                  id
+                  priceCents
+                  artworkId
+                  editionSetId
+                  quantity
+                }
+              }
+            }
+          }
+          errors
+        }
+      }
+    `
+      return graphql(exchangeSchema, mutation, null, context, {
+        orderId,
+        fulfillment,
+      }).then(result => {
+        const {
+          order,
+          errors,
+        } = result.data.ecommerce_fulfillWithOneFulfillment
+        return {
+          order,
+          errors,
+        }
+      })
+    },
+  }
+)
